Extract tests for run_plan::report to new file.
